+++ src/main/kotlin/by/course/govservices/dto/PaymentDto.kt
@@ -4,9 +4,9 @@ import java.math.BigDecimal
 import java.time.LocalDate
 
 data class PaymentDto(
-    val id: Int?,
-    val bidId: Int,
+    val id: Long?,
+    val bidId: Long,
     val sum: BigDecimal,
     val date: LocalDate,
-    val statusId: Int
+    val statusId: Long
 )
